Fix compile on mac gcc, since it's broken in its handling of static const int class members, solution is to make a copy of the member before using it! Fortunately I am elite and know surrounding the reference in int() makes a copy.
git-svn-id: svn+ssh://svn.last.fm/svn/misc/max/v2@85818 ab8f4a45-97f9-0310-bbd1-854ce3dcee89

+++ src/lib/moose/TrackInfo.cpp
@@ -227,14 +227,14 @@ TrackInfo::scrobblePoint() const
 {
     // If we don't have a length or it's less than the minimum, return the
     // threshold
-    if ( duration() <= 0 || duration() < kScrobbleMinLength )
+    if ( duration() <= 0 || duration() < int(kScrobbleMinLength) )
         return kScrobbleTimeMax;
 
-    float scrobPoint = qBound( kScrobblePointMin,
+    float scrobPoint = qBound( int(kScrobblePointMin),
                                Moose::Settings().scrobblePoint(),
-                               kScrobblePointMax );
+                               int(kScrobblePointMax) );
     scrobPoint /= 100.0f;
 
-    return qMin( kScrobbleTimeMax, int( duration() * scrobPoint ) );
+    return qMin( int(kScrobbleTimeMax), int( duration() * scrobPoint ) );
 }
